Replace tests to android test and make them work again
