+++ app/src/androidTest/java/nl/tudelft/cs4160/trustchain_android/block/TrustChainBlockTest.java
@@ -1,4 +1,4 @@
-package nl.tudelft.cs4160.trustchain_android.Block;
+package nl.tudelft.cs4160.trustchain_android.block;
 
 
 import android.support.test.runner.AndroidJUnit4;
@@ -10,7 +10,6 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.libsodium.jni.NaCl;
 
-import nl.tudelft.cs4160.trustchain_android.block.TrustChainBlockHelper;
 import nl.tudelft.cs4160.trustchain_android.crypto.DualSecret;
 import nl.tudelft.cs4160.trustchain_android.crypto.Key;
 import nl.tudelft.cs4160.trustchain_android.main.OverviewConnectionsActivity;
